funktiot omaan tiedostoon
